LoadLibrary: Redirect system32 loads to injection directory
If we are loading via injection from outside of the game directory we
need to redirect library loads from system32 to our directory explicitly
rather than just stripping off the directory.

However, if the load is for nvapi and we have not shipped our wrapper
this redirect will fail, so still fall back to just using the library
name.

This gets our nvapi wrapper loading in RE2 via injection from outside of
the game directory.