Injector: Warn if module loaded from both game & injection dirs
If we inject 3DMigoto from an external directory and there is also a
second copy of 3DMigoto located in the game directory the game will
crash. I haven't looked into exactly why this is, but it's not something
we would want to do anyway, so have the loader detect this situation and
throw up a big warning asking the user to remove one.

Running the injector from inside the game directory works fine, since
the module it is trying to inject is the same as the module the game
will be trying to load, so it doesn't get double loaded.