Add a "debug_locks" lock dependency checker
This adds a locking dependency checker that can detect potential
deadlock hazards between 3DMigoto, the game, DirectX and the driver.
This is inspired by lockdep from the Linux Kernel, but is not derived
from it (wrong OS, wrong locking infrastructure, no interrupt disabled
code, small enough problem space not to need locking classes).

It works by hooking into the CriticalSection routines to allow it to
build a graph of dependencies between all locks (including those used by
DirectX, the driver and the game) and look for scenarios where locks
were taken in a different order (e.g. lock B was taken while holding
lock A this one time, but then this other time lock A was taken while
holding lock B), which indicates a potential deadlock hazard if they
ever happen to coincide.

This means like lockdep it doesn't need to actually hit a deadlock to
find a hazard - it just needs to see both sides of the code involved in
the deadlock run to determine if there is a possible hazard or not.

This doesn't keep track of past stacks, so unlike lockdep when it
detects a hazard it can only dump the lock stack of the current thread,
which is only one side of the deadlock condition. When a hazard is
detected it wipes it's record of other stacks it has checked so that
(assuming we aren't actually about to hit the deadlock) when the other
side repeats the actions that lead to the hazard it will be reported
then. This does mean that we will sometimes miss the other side of the
deadlock (e.g. if the code is never run again or we are actually about
to hit the deadlock), but we can hopefully just repeat a few times to
get a good run, and reloading config may help cause 3DMigoto to re-run
code on our side.

Like lockdep it tracks the current lock stack (though does not warn if
locks are dropped out of order, because it seems that happens routinely
in DirectX), maintaining a hash that it uses to determine if it has
already checked the current locking stack so as not to kill performance.

For example, it can locate locking bugs between DirectX and 3DMigoto
such as the following:

    40ec: Potential deadlock scenario detected: Lock 000001F0F60F17D0 taken after 00007FFF591AF100
    40ec held locks (most recent first):
    40ec: EnterCriticalSection(000001F0F60F17D0) C:\WINDOWS\SYSTEM32\d3d11.dll+0x26649
    40ec: EnterCriticalSection(00007FFF591AF100) C:\cygwin64\home\dss\3d-fixes\DOA6\d3d11.dll+0xc69dc
    40ec: EnterCriticalSection(000001F0ED9FCC30) T:\SteamRU\steamapps\common\Dead or Alive 6\DOA6.exe+0x96af4f
    40ec: EnterCriticalSection(000001F0EDA167D0) T:\SteamRU\steamapps\common\Dead or Alive 6\DOA6.exe+0x62b2a5
    ...
    40ec: Potential deadlock scenario detected: Lock 00007FFF591AF100 taken after 000001F0F60F17D0
    40ec: Potential deadlock scenario detected: Lock 00007FFF591AF100 taken after 000001F0F60F17D0
    40ec held locks (most recent first):
    40ec: EnterCriticalSection(00007FFF591AF100) C:\cygwin64\home\dss\3d-fixes\DOA6\d3d11.dll+0x110290
    40ec: EnterCriticalSection(000001F0F60F17D0) C:\WINDOWS\SYSTEM32\d3d11.dll+0x2792e
    40ec: EnterCriticalSection(000001F0F60F17D0) C:\WINDOWS\SYSTEM32\d3d11.dll+0x26649
    40ec: EnterCriticalSection(000001F0EDA167D0) T:\SteamRU\steamapps\common\Dead or Alive 6\DOA6.exe+0x62b2a5

This is showing that 3DMigoto called into DirectX while holding our
lock, while at another point in time DirectX called into 3DMigoto while
holding their lock, and if these both happen to coincide on different
threads we will get a deadlock.

This highlights quite a lot of potential deadlocks we may need to
investigate. Not all of them involve 3DMigoto - quite a few of them
appear to be between the driver, OS and/or the game, and it seems like
most of them are reported on the same thread (which may help side step
any issues that are present), but they may be of interest in thier own
right.