Fix AB-BA deadlock between ShaderRegex and resource release tracker
In DOA6 we sometimes get hangs when loading stages, which is the result
of an AB-BA type deadlock between the DirectX calling the resource
release tracker with its lock held, and ShaderRegex calling into DirectX
with our lock held. This shows up in the lock debugger as follows:

    3640: Potential deadlock scenario detected: Lock 000002B822916AA0 taken after &G->mCriticalSection
    3640 held locks (most recent first):
    3640: EnterCriticalSection(000002B822916AA0) C:\WINDOWS\SYSTEM32\d3d11.dll+0x5b161
    3640: EnterCriticalSection(&G->mCriticalSection) HackerContext::DeferredShaderReplacementBeforeDraw:617
    ...
    4b88: Potential deadlock scenario detected: Lock &G->mCriticalSection taken after 000002B822916AA0
    4b88 held locks (most recent first):
    4b88: EnterCriticalSection(&G->mCriticalSection) ResourceReleaseTracker::Release:1304
    4b88: EnterCriticalSection(000002B822916AA0) C:\WINDOWS\SYSTEM32\d3d11.dll+0x247f9
    4b88: EnterCriticalSection(000002B8890A0130) T:\SteamRU\steamapps\common\Dead or Alive 6\DOA6.exe+0x968015

To fix this deadlock we reduce the length of time that we hold our
critical section so that it protects our updates to the
OriginalShaderInfo entry, but is not held when we call into DirectX to
bind the replacement shader.