Lock debugger: Ignore locks in ntdll.dll
ntdll.dll has a lock that seems to be taken all over the place -
sometimes it is held while 3DMigoto/game/DirectX is called, and other
times it is taken while 3DMigoto/game/DirectX calls into ntdll.

I don't know for sure what this lock is - we've never seen it involved
in an actual deadlock and while the locking stacks do look alarming I
suspect that there is something else going on that will turn out to make
it harmless (at a guess, perhaps it is some inner workings of Windows'
context switch mechanism).

Until proven otherwise, treat it as a false positive and ignore all
locks taken/released by ntdll.dll. This change results in the only
warnings we now see in DOA6 representing actual locking bugs.