Fix deadlock between ShaderUsage and resource release tracker
This fixes the following AB-BA deadlock between stat collection for the
ShaderUsage.txt and the resource release tracker:

    > Hunting toggled to 1
    0c8c: Potential deadlock scenario detected: Lock 000001231D9E4C90 taken after &G->mCriticalSection
    0c8c held locks (most recent first):
    0c8c: EnterCriticalSection(000001231D9E4C90) C:\WINDOWS\SYSTEM32\d3d11.dll+0x5b161
    0c8c: EnterCriticalSection(&G->mCriticalSection) HackerContext::BeforeDraw(709)
    ----- Previously seen locking pattern that could lead to an AB-BA deadlock:
          EnterCriticalSection(&G->mCriticalSection) ResourceReleaseTracker::Release(1304)
          EnterCriticalSection(000001231D9E4C90) C:\WINDOWS\SYSTEM32\d3d11.dll+0x2792e
          EnterCriticalSection(000001231D9E4C90) C:\WINDOWS\SYSTEM32\dxgi.dll+0x65b8
          EnterCriticalSection(0000012315DB67D0) T:\SteamRU\steamapps\common\Dead or Alive 6\DOA6.exe+0x62b2a5

The code is reorganised to allow the various DirectX APIs to get state
to be called before taking the critical section to protect our usage
data structures from simultaneous modification.

With this change, the lock checker does not locate any further deadlocks
during ordinary gameplay of DOA6.