Lock debugger: Dump call stack when locking issue found
The lock stack is useful to track down many locking issues, but if the
lock is held over a very large section of code (as it is in our config
reload) we may need to narrow things down further. This dumps the
current thread's call stack when a locking issue is detected, which can
help point the finger at the specific culprit.

This currently only dumps the current thread's call stack and not the
call stack on the other side of the locking issue, so we might want to
extend it later to do that if necessary.

Also add a __debugbreak() if a debugger is attached so we can
immediately examine the current state from there.