Add a dedicated lock to protect mResources
This splits mResources' protection out of mCriticalSection into its own
lock, thereby removing the restriction that we cannot have
mCriticalSection held while calling DirectX and replacing it with the
restrictions (debug_locks=1 can help identify violations of these):

- We cannot call into DirectX while holding mResourcesLock. Failure to
  follow this rule can lead to a 2-way deadlock.

- If mResourcesLock and mCriticalSection ever need to be obtained
  simultaneously, mCriticalSection must be obtained first. Failure to do
  so can lead to a 3-way deadlock.

I was initially concerned this would be a significant and hard to verify
change, but since we can safely hold mCriticalSection and mResourcesLock
simultaneously if we follow the above rules we can continue using
mCriticalSection for all other structures and only take mResourcesLock
briefly when we need it - this is essentially an iterative move away
from the one global lock model.

This removes some of the various AB-BA deadlock warnings in the code
since this change makes most of them no longer relevant. The warning
next to the resource release tracker is left in place, and a new warning
is added at the declaration of the lock.