Patch disassembly resource bindings to look like d3dcompiler_46
This adds a d3dcompiler_46 backwards compatibility fixup pass to the
disassembler to reformat the resource bindings section to match the
output from the older disassembler changing the HLSL bind field back
into a slot field and matching the table alignment of the old
disassembler.

This should allow ShaderRegex patterns matching these lines to continue
working when we switch to the vs2017 branch without needing any
modifications.