vs2017: Patch disassembly resource bindings to look like d3dcompiler_46
This is from the vs2017 branch, included in master to keep the diff
between the two branches minimal. In this branch the functionality is
disabled by default in util.h and the command line parameter to enable
it in d3d_compiler is omitted. It wouldn't hurt to enable it, but it is
unnecessary work that we don't need until we switch master to vs2017.

Original commit message:

This adds a d3dcompiler_46 backwards compatibility fixup pass to the
disassembler to reformat the resource bindings section to match the
output from the older disassembler changing the HLSL bind field back
into a slot field and matching the table alignment of the old
disassembler.

This should allow ShaderRegex patterns matching these lines to continue
working when we switch to the vs2017 branch without needing any
modifications.