Dis/Assembler: Handle "undecipherable custom data"
Currently our disassembler fixup routine strips undecipherable custom
data tags, and the assembler throws a parse error if it encounters one
(such as, from an old shader dumped prior to the addition of the fixup
routine).

This adds support for these tags to both - when the disassembler fixup
routine encounters one it now properly decodes the length and appends
the raw data to the line in a hexdump fashion. When the assembler
encounters one of these tags it will then reassemble the hexdump back
into bytecode, and if there is no hexdump (from an old shader) it will
skip over the instruction without assembling anything.

Most of our disassemble routines will now include custom data by
default to make sure that any tags that could cause problems are visible
(if the shaderhacker wants to strip them out that's fine, but I'd rather
they see the tag exists and make a conscious decision to that end), but
for the sake of backwards compatibility with any ShaderRegex patterns
that might be sensitive to one of these tags suddenly appearing it must
be enabled in the d3dx.ini for ShaderRegex to include it.

With this change the Project Cars example shaders are able to
reconstruct for the HLSL tests again, and the assembly tests are
passing.