Fix regression assigning resources when ReShade is present
The commit (and related commits):

    commit 385991c84113958207026a239481cf2f27440909
    Author: Ian Munsie <darkstarsword@gmail.com>
    Date:   Mon Feb 11 18:18:39 2019 +1100

        Resource Pool: Discard resources from wrong devices

Introduced a regression when 3DMigoto and ReShade were used together,
and [Resource] assignments were in use. When ReShade is in use the
resource->GetDevice() will return a different value to the device we
have wrapped (presumably GetDevice() will return the real DX device, and
the one we have wrapped is the ReShade device), causing us to mistakenly
think that a device swap has taken place and invoking our inter-device
resource code paths.

This leads to a slow down, and if this code path is invoked on a
resource that is not yet supported by the inter-device resource copy
(such as a dynamic resource) the resource will be discarded, resulting
in black objects.

This issue was reported on DOAXVV when new versions of both 3DMigoto and
ReShade were in use simultaneously, and reproduced using DOA6 and
ReShade 4.2.1.

To fix this, we no longer look up the device that created the resource
from the resource itself, and instead store it in the custom Resource
structure. This is a bit of a pain because it requires keeping the two
in sync and everywhere the resource may be changed we need to update the
device, but testing in DOAXVV (after reverting the commit to disable
re-running [Constants] on new device creation and trying one of the
crashy mods while opening the news) it seems to do the job.