ShaderRegex: Prevent linking the duplicate command lists
If the game created a shader with the same hash as an existing shader
(whether due to being identical shaders, or the same shader that had
been released and recreated) ShaderRegex would link the command list
once for each creation. This could lead to the command lists growing
over time, costing more and more performance until the config was
reloaded.

To address this, the ShaderRegex will now keep a reference to the link
command, and check if it has already been added to a ShaderOverride's
command list before adding it again. It now adds the same link command
to each matching ShaderOverride, so the references to the matched shader
have been removed from the ini_line and optional generated section.