Command List: Allow indirect draw call buffer to be assigned
This makes the "this" target assignable for indirect draw calls, which
gives us a way to force the indirect buffer to mono at the last moment,
e.g.

	[ShaderOverrideIndirectParticles]
	hash = ...
	this = mono copy this
	vb0 = mono copy vb0
	ib = mono copy ib

It can be better to identify the exact buffers at fault and force them
to mono, but if the buffer hashes vary and can't be locked down with
fuzzy texture overrides this provides an alternative.