Turn draw command arguments into expressions
This replaces the integer arguments to the various draw & dispatch calls
with expressions, allowing for more flexibility when calling these, e.g.
one could fire up a compute shader with thread groups for each 16x16
block of a resource via (note: rt_width/height operates on o0):

    dispatch = (rt_width + 15) / 16, (rt_height + 15) / 16, 1

This could also give more flexibility with modifying a draw call from
the game where some arguments are to be overridden, but not others, e.g.
if a vertex buffer had been altered and had it's base vertex offset
lopped off:

   drawindexed = index_count, first_index, 0
   handling = skip

or altering the buffer of an indirect draw call but leaving the offset
intact:

   [ResourceMonoThis]
   [ShaderOverrideIndirect]
   hash = ...
   ResourceMonoThis = mono copy this
   DrawIndexedInstancedIndirect = ResourceMonoThis, indirect_offset
   handling = skip