Command list: Log errors creating destination resources to frame analysis
Most of the command list logs to the frame analysis log, but this
particular error message went to the debug log, so the error was not
apparent if only looking at the frame analysis log.

This redirects the message to the frame analysis log to see it in
context, though the specific resource creation failure is left in the
regular log since that is generally not re-attempted each frame.