Custom Resources: Allow misc_flags to be overridden
This propagates misc_flags in much the same way as we already do for
bind_flags and allows it to be overridden by the user to set an
arbitrary value if desired.

It automatically populates the drawindirect_args flag for resources used
in indirect draw calls, but otherwise this is mostly to facilitate
experimentation.