Clear raw misc flag when neither SRV or UAV bind flags are set
DOA6 has the raw flag set for the buffer used with indirect draw calls,
but if we create a mono copy of that buffer to fix rendering issues we
lose the SRV and UAV bind flags that it also has since we aren't binding
the copy to any shaders. Turns out that having the raw flag set without
either of those bind flags is another one of those weird undocumented
edge cases where DirectX refuses to create a resource.

Detect this situation and clear the raw flag.