nvapi: Do not wrap non-supported nvapi calls
In the event that nvapi_QueryInterface returns NULL for a given stereo
call we would still implement that call ourselves, which could lead to a
NULL pointer dereference when we go to call the original function.

Given that nvidia plans to drop 3D Vision support past R418 there is a
real possibility that some of the stereo calls may be dropped and we may
well end up in this situation. Pre-emptively disable wrapping any calls
that nvapi_QueryInterface returns NULL for now to avoid possible crashes
tomorrow.