Loader: Set working directory if launch= specifies a path
If a user sets launch= to the game's executable the loader will attempt
to launch that, but will leave the working directory set as the loader's
directory, which can confuse some games.

This deduces the working directory that should be used from the path
specified in launch, checking that the target exists and lopping off the
file part if necessary.

If the steam browser protocol is in use or it otherwise cannot deduce
the working directory it will pass NULL for the working directory,
leaving that set to the loader directory as before.