Add 'condition = stereo_available' to included ini
This adds a preamble section to the ini file above any section, that can
include a "condition" line determining whether the rest of the ini file
is to be included or not. This condition is evaluated via the command
list expression evaluator, but it must be statically evaluatable.

A statically evaluatable "stereo_available" is added that returns
whether 3D Vision is enabled in the control panel (differing from
"stereo_active" which can change at runtime and is only sometimes
statically evaluatable), thereby allowing included ini files that depend
on 3D Vision to be entirely omitted when 3D Vision is unavailable.