Allow included ini files to override their namespace
The ini file namespaces defaults to their filenames for both
convenience, and avoiding too many mods copy & pasting a namespace that
is not theirs. This works, but it means if anything inside an ini file
is to be accessed from elsewhere it must always be placed in the same
path, which users don't always manage to get right.

This adds a namespace setting to the ini file preamble, which can be
used to override the namespace for the ini, allowing ini files that
declare variables or resources intended to be used in other ini files to
have a well defined namespace regardless of exactly where the ini file
is located.

Custom shader and resource files referred to from the ini files are
still searched for relative to the ini path rather than the namespace.