Add filter_index to [ShaderOverride] and [ShaderRegex]
This adds a new form of partner shader filtering that works in much the
same way as texture filtering - a filter_index can be set in a
[ShaderOverride] or [ShaderRegex] section and tested via "variable =
shader_type" (e.g. x=ps) from the partner shader.

This form also allows the shader slot to be tested to see if a shader is
bound at all, which can be useful to filter out depth only passes
(wanted for DOA6 modding), and doesn't rely on assigning variables from
command lists in pipeline order.