Divide constant buffer offsets by 16
The constant buffer offsets in the disassembly are byte offsets, but
that's not super useful and we end up having to divide / multiply by 16
to match it up with the constant buffer indices used in the code. This
is even more problematic for matching a use of a particular entry with
ShaderRegex since that can't divide by 16, and while it is possible to
use the offset in code (via udiv and a dynamic calculated index at
runtime), it is inconvenient.

This adds an option to perform a fixup while disassembling shaders to
replace the offsets with the constant buffer index, components and/or
index range. It is disabled by default for compatibility with existing
ShaderRegex patterns, but shipped enabled in the template as it is
anticipated to be useful.

Variables that fit within a single index will have the offset replaced with
an index and components, and the Size is replaced with a Component count:

    //   float localIblMipmapBias;          // Offset:   80 Size:     4 [unused]
    //   float screenAspectRatio;           // Offset:   84 Size:     4 [unused]
    //   float2 invResolution;              // Offset:   88 Size:     8
    //   float4 shadowMapSizeAndInvSize;    // Offset:   96 Size:    16 [unused]
                                                      -- vv --
    //   float localIblMipmapBias;          // Index:    5.x              Components:     1 [unused]
    //   float screenAspectRatio;           // Index:    5.y              Components:     1 [unused]
    //   float2 invResolution;              // Index:    5.zw             Components:     2
    //   float4 shadowMapSizeAndInvSize;    // Index:    6.xyzw           Components:     4 [unused]

Matrices or other items occupying up to four indices are expanded like this
to allow for ShaderRegex to easily match all indices:

    //   row_major float4x4 g_mViewProj;    // Offset:    0 Size:    64 [unused]
                                                      -- vv --
    //   row_major float4x4 g_mViewProj;    // Index:    0 1 2 3          Components:    16 [unused]

Items occupying more than four indices are expanded into a range as a
compromise between clarity and ShaderRegex matches (can still match
first/last in the pattern, and patching the shader to use dynamic indexing
can allow other indices to be used in code):

    //   row_major float4x4 g_mLightProj[4];// Offset:  128 Size:   256 [unused]
                                                      -- vv --
    //   row_major float4x4 g_mLightProj[4];// Index:    8-23             Components:    64 [unused]

Items occupying multiple indices, but where the last index is only partially
occupied are also expanded to ranges, with the final component shown:

    //   float _NeighborIsAdapting[4];      // Offset:   96 Size:    52
                                                      -- vv --
    //   float _NeighborIsAdapting[4];      // Index:    6-9.x            Components:    13

Entries in embedded structures that lack the "Size" field cannot be
processed as richly as straight constant buffer entries since we can't
(trivially) calculate their end index/component, but their Offset is still
turned into an Index with a component shown if it is not aligned to x:

    //       int m_PatchX;                  // Offset:    0
    //       int m_PatchZ;                  // Offset:    4
                                                      -- vv --
    //       int m_PatchX;                  // Index:    0
    //       int m_PatchZ;                  // Index:    0.y