Drop support for zero marking mode
The zero marking mode has always been buggy, has long since bit-rotted,
just tends to crash if anyone tries to use it nowadays, is largely
unnecessary in light of the pink marking mode, and adds a bunch of
complexity to our shader replace routines we could do without.

If someone wants to try to restore it to a working condition they are
more than welcome to, but for now: Drop it.