Decompiler: Refactor parameter parsing
This moves all the random auto-fix parameters the decompiler takes into
their own structure and just passes a pointer to that structure instead
of the eye sores / manual mem copies we have at the moment.

We might consider dropping all these auto-fixes, but I would like to
look at them a little closer to see if they are salvageable and make
sure no one is using them before committing to that.