nvapi: Work around RE2+DMC5 SetDepthBoundsTest game bug
The fix for NvAPI_D3D11_SetDepthBoundsTest didn't make a whole lot of
sense, because it seemed to just be calling through to the original
nvapi call with the only functional changes being a null pointer check
and returning NVAPI_OK.

Adding some debugging revealed the true cause of the issue:

    NvAPI_D3D11_SetDepthBoundsTest(0000000008518E10, 1, 1.609178, 1.609178) -> -5

Every now and then the game passes fMinDepth == fMaxDepth != 1.0, which
causes nvapi to return NVAPI_INVALID_ARGUMENT and in turn triggers what
appears to be a game bug which stops rendering any further geometry.

This adds an explicit test for this case and overrides the return value
to NVAPI_OK, resolving the issue.

Also removes the redundant nvapi_QueryInterface call in favour of using
the pointer saved during the original nvapi_QueryInterface for
SetDepthBoundsTest, and NULL pointer checks that nvapi is capable of
doing themselves.