ShaderRegex: Add caching
This adds a cache to ShaderRegex to reduce stutters when the game starts
using a new shader during gameplay. If caching is enabled (using the
existing cache_shaders setting) then the ShaderRegex engine will write
out the patched shader binary and some metadata to ShaderCache (keeping
ShaderFixes for manually edited shaders), and will use that to skip
processing the shader next time it is loaded. The filenames have _regex
to distinguish them from dumped shaders.

To determine if the cached data is valid, the [ShaderRegex] ini sections
are hashed along with any other settings that may affect the disassembly
output. This hash is written into the cache metadata along with a
metadata file version and compared when the shader is next loaded. If
the hash does not match the shader will be run through ShaderRegex
analysis again.

In addition to the hash and file version, the metadata also indicates
whether the shader was patched (indicating the presence of a saved .bin
file), and includes indices of all the ShaderRegex groups that matched
that shader.

Since the ShaderRegex groups are sorted if the hash matches than the
groups should be in the same order as they were when the cache was
created and the indices can be used to quickly look up the matching
groups without having to do a string lookup - a vector of pointers to
the groups is saved to allow for random access. Then the group's command
lists, filter_index and tag line can be linked in to the ShaderOverride.

The metadata is saved even if ShaderRegex does not match the shader
(matches=0, patched=false) - this allows the analysis to be skipped next
time.