Add "regex" marking action to dump ShaderRegex result
This adds a new "regex" marking action that can be used to output the
result of ShaderRegex on a shader.

We don't have any good way to save off the patched assembly or bytecode
for these (notably the patched bytecode would be missing RDEF), so run
ShaderRegex over the original bytecode and dump the result of that.

Noting that ShaderRegex may have linked a command list into the
ShaderOverride and/or set a filter_index that would no longer be applied
on the next config reload we unlink any ShaderRegex command lists and
restore a backup of filter_index when dumping these.

This will hopefully make the need to log the shader assembly obsolete,
so remove that from the log.