ShaderRegex: Explicitly log cached misses
We cache the result of ShaderRegex even when it doesn't match any
patterns so that we can skip doing the analysis next time, and that just
works - but the log message suggests that a command list was loaded from
cache, when in actuality nothing was loaded from cache.

This adds an explicit return value for the NO_MATCH case to log
appropriately, and renames INVALID to NO_CACHE to make the code clearer
that return value means that analysis must be applied.