Fix mistaking operand for shader target match on VS2013
The "BUG: Unknown shader filter type" dire error was being triggered in
DOA6 when 3DMigoto was built with the vs2013 toolchain, but not when
built with the vs2017 toolchain. Some operands (such as "window_width"
from the help text shader) were being mistaken for shader targets by the
parser, triggering the BUG warning because 'w' is not a valid shader
type.

The parsing for these uses scanf("%lcs%n") with %lc to match the shader
type and %n to verify that the entire string was matched. However, if
the match stops after %lc and before %n it will not be filled out, and
because it doesn't increment the return value there is no immediate
indication that this has happened. In vs2013 it is left uninitialised
and if the stack garbage it occupied happened to contain the string
length the test would incorrectly pass. This does not appear to be the
case in vs2017 - either it always writes a value to %n, or we got lucky
and the stack garbage did not happen to hold the string length.

This initialises the variable used to store %n to make sure that the
test will not give a false positive. It also checks to make sure the
shader type is valid as we do in the texture target code paths.