Upscaling: Remove requirement to pass DXGIFactory
One of the upscaling modes involves creating a second swap chain from
the IDXGIFactory, which is currently passed from the CreateSwapChain
method. In a following patch we want to be able to call wrap_swap_chain
from D3D11CreateDeviceAndSwapChain without the burden of needing to find
the DXGIFactory used to create the swap chain since we don't have it
handed to us in that call.

Remove the requirement to pass the DXGIFactory to the upscaling swap
chain and have it locate this itself if it needs it.