Fix compatibility issue with Steam Overlay + controller
In some situations the Steam overlay and controller will inconsistently
not work with 3DMigoto. The flow that leads to this happening is:

1. Steam and 3DMigoto both hook DXGI in a random unpredictable order.
   The fail case is specifically when 3DMigoto hooks in prior to Steam.
2. Game calls CreateDeviceAndSwapChain(), which goes into 3DMigoto
3. 3DMigoto reimplements this call in terms of separate CreateDevice
   (uninteresting) and CreateSwapChain calls
4. 3DMigoto employs multiple (and in some cases redundant) strategies to
   avoid games of hot potato that we have previously hit with other
   third party tools, one of which is to directly call an internal
   UnhookableCreateSwapChain as opposed to calling the hooked (by both
   Steam and 3DMigoto) DXGIFactory::CreateSwapChain
5. UnhookableCreateSwapChain calls straight through to the trampoline /
   original CreateSwapChain. Depending on whether Steam or 3DMigoto got
   in first this might then call into Steam's hook, or it might not (if
   Steam hooked first it will work fine, if Steam hooked in second it
   will fail).

In order to resolve this issue, this commit partially reverts an earlier
change made to CreateDeviceAndSwapChain. The change in question (commit
titled "Break CreateDeviceAndSwapChain into CreateDevice and
CreateSwapChain.", sha d9560cf70e66555155ce0e6188e83c6f63998c03) aimed
to simplify our init paths and eliminate future copy & paste errors by
effectively removing one possible init path. This unfortunately had the
downside that it would break the expectations of other tools that might
also be trying to hook into the game. One such example that we already
dealt with was SpecialK, which performed a similar init path
simplification by redirecting CreateDevice to CreateDeviceAndSwapChain
with a NULL swap chain, and the two interacted in such a way that we
ended up in a game of hot potato until the stack ran dry and crashed.

This issue with the Steam overlay is another case of the altered init
paths breaking their expectations, and restoring the original flow of
CreateDeviceAndSwapChain will allow their hook to work as they expect.
This reversion also makes the prior unhookable routine workaround for
SpecialK redundant, and that may be removed in a future commit.

In the time since that change was originally introduced refactoring has
been done on the device and swap chain creation calls (see:
"HackerCreateDevice: Split out device & context wrapping into separate
function" and "CreateSwapChain: Split up into several functions") to
move common code into utility routines so that the reverted code is
now substantially simpler than it was in the past.

As a side note, it is worth noting that Steam does not hook into any of
the device creation functions, but rather only hooks the DXGIFactory*
CreateSwapChain methods. When those are called from the game this is
fairly obvious why it works, but in the D3D11CreateDeviceAndSwapChain
path they are actually relying on an internal implementation detail of
DirectX where DirectX will call that exported method. If Microsoft were
ever to change that internal behaviour (say, by directly calling an
internal function that is not exposed publicly, similar to our
unhookable routines) Steam's overlay would break. Steam does hook into
the device from CreateSwapChain, but as far as I can tell they only hook
the Release() call, since their main interest is in the swap chain's
Present() call and buffer / full screen management.

Closes #112