Upscaling: Remove use of Deviare pass through trampoline
Upscaling calls the fnOrigCreateSwapChain function pointer to call
through the original CreateSwapChain routine, but that pointer will only
be initialised if 3DMigoto has had an opportunity to hook the routine
from our CreateDXGIFactory hook. We currently force this in our init
paths, but that is far flung from this use case and would be too easy to
refactor out without noting this depends on it in some cases.

Therefore, call the pFactory->CreateSwapChain instead, since we already
have the pointer to do that with, and our hooking reentrancy quirk
detection will already prevent 3DMigoto doing anything with that call.

Remove the header declaration of fnOrigCreateSwapChain and make it
static to restrict references to it local to a single file.