Add beginnings of a custom shader #include handler
This is the beginnings of a custom include handler that will eventually
be able to keep track of what files each HLSL shader depends on, so that
it can be reloaded if one of its dependencies gets updated but it itself
did not.

For now however this is just a first pass at implementing a custom
include handler aimed at being a drop in for the standard include
handler with no behaviour changes. Since the various edge cases of the
standard include handler are not exactly documented anywhere we don't
exactly have a checklist to determine if that goal has been met and this
commit will serve as a comparison point.

Known edge cases that we do handle:
- Both / and \ may be used as path separators