Include handler: Fall back to including from the CWD
The custom include handler needs to match the behaviour of the standard
include handler to maintain backwards compatibility. If the standard
include handler cannot locate a file relative to the file that included
it, it will fall back to searching from the current working directory -
we therefore have to do the same.