Include handler: Drop directory tracking for backwards compatibility
It turns out that the standard include handler is actually pretty dumb
when recursively including files. Rather than included files being able
to include other files relative to themselves, it always includes files
relative to the *initial* source file or current working directory.

For compatibility with the standard include handler this drops the use
of the directory stack used to enable sensible recursive file inclusion.
However, since this is probably something we do want, the code is left
largely intact so it can be trivially enabled later if we decide that
enabling it is worth breaking behaviour in some edge cases, or add a
config parameter to select it.