Add recursive_include option to enable sensible include handling
As noted in the previous commit, the standard include handler is rather
silly in how it handles files in subdirectories that include other
files, with it searching for the files relative to the initial source
file rather than the file actually issuing the #include statement.

We can handle this in a much more sensible manner, but since there are
some (admittedly contrived) edge cases where doing so could change
behaviour add an option to select the new sensible behaviour.

While we are at it, make this option also disable some other silly
behaviour of the standard include handler - normally if it cannot find
an included file relative to the initial source file it will search from
the current working directory instead. This can be problematic, since
some games have inconsistent working directories (e.g. Far Cry 4 &
Primal use different working directories depending on whether they are
they Steam or UPlay editions) and so accidentally using #include
<ShaderFixes/hud.hlsl> (i.e. relative to the working directory) would
only work on one edition of the game. Make this option disable searching
from the working directory to ensure that the correct canonical path is
always specified.

Finally, add an undocumented -1 mode to this option to switch back to
the standard include handler to ease testing potential differences
between the two.