Include Handler: Add support for unicode filenames
The D3DCompiler API uses UTF8 - it's *almost* like Microsoft worked out
their big mistake in choosing the poo-brain worst-of-all-worlds
considered-harmful UTF16... almost. Unfortunately CreateFileA, like many
Windows APIs (and unlike any other OS' APIs) doesn't accept UTF8
filenames, so we have to use the CreateFileW variant to allow it to open
files with unicode characters, meaning we have to convert the UTF8
string back to UTF16... after we spent all that effort converting it to
UTF8 to pass to D3DCompile in the first place.